reorganize new project service definitions
